requirements.txt
streamlit==1.29.0
psycopg2-binary==2.9.9
opencv-python-headless==4.8.1.78
qrcode[pil]==7.4.2
pandas==2.1.3
numpy==1.26.2